Add confirm message before delete

+++ src/components/CustomerList.js
@@ -2,7 +2,8 @@ import React, { Component } from "react";
 import ReactTable from "react-table";
 import "react-table/react-table.css";
 import Button from "@material-ui/core/Button";
-
+import AddCustomer from "./AddCustomer";
+import { Snackbar } from "@material-ui/core";
 export default class componentName extends Component {
   constructor(props) {
     super(props);
@@ -19,13 +20,33 @@ export default class componentName extends Component {
   };
 
   deleteCustomer = customerLink => {
-    fetch(customerLink.original.links[0].href, { method: "DELETE" })
-      .then(this.loadCustomers())
+    if (window.confirm("Are you sure?")) {
+      fetch(customerLink.original.links[0].href, { method: "DELETE" })
+        .then(res => this.loadCustomers())
+        .then(res => this.setState({ open: true, message: "Customer deleted" }))
+        .catch(err => console.error(err));
+    }
+    // fetch(customerLink.original.links[0].href, { method: "DELETE" })
+    //   .then(this.loadCustomers())
 
-      .catch(err => console.error(err));
-    console.log(customerLink.original.links[0].href);
+    //   .catch(err => console.error(err));
+    // console.log(customerLink.original.links[0].href);
   };
-
+  saveCustomer = customer => {
+    fetch("https://customerrest.herokuapp.com/api/customers", {
+      method: "POST",
+      headers: {
+        "Content-Type": "application/json"
+      },
+      body: JSON.stringify(customer)
+    })
+      .then(res => this.loadCustomers())
+      .then(res => this.setState({ open: true }))
+      .catch(err => console.error(err));
+  };
+  handleClose = () => {
+    this.setState({ open: false });
+  };
   render() {
     const columns = [
       {
@@ -54,10 +75,24 @@ export default class componentName extends Component {
     ];
     return (
       <div>
+        <AddCustomer saveCustomer={this.saveCustomer} />
         <ReactTable
           data={this.state.customers}
           columns={columns}
           filterable={true}
+        />
+        <Snackbar
+          anchorOrigin={{
+            vertical: "bottom",
+            horizontal: "left"
+          }}
+          open={this.state.open}
+          autoHideDuration={3000}
+          onClose={this.handleClose}
+          ContentProps={{
+            "aria-describedby": "message-id"
+          }}
+          message="Customer added successfully"
         />
       </div>
     );
